requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Markdown==3.4.3
WTForms==3.0.1
email-validator==2.0.0.post1
python-dotenv==1.0.0
